move candidate
