+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/stratocumulusy.iml" filepath="$PROJECT_DIR$/.idea/stratocumulusy.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/Stratocumulusy.iml" filepath="$PROJECT_DIR$/.idea/Stratocumulusy.iml" />
     </modules>
   </component>
 </project>
