+++ src/Routes.js
@@ -12,7 +12,8 @@ import Candidate from "./containers/Candidate";
 import TestList from "./containers/TestList";
 import Results from "./containers/Results";
 import SolveTest from "./containers/SolveTest";
-
+import AddCandidate from "./containers/AddCandidate";
+import ConfirmationCode from "./containers/ConfirmationCode";
 export default function Routes({ appProps }) {
     return (
       <Switch>
@@ -25,6 +26,8 @@ export default function Routes({ appProps }) {
         <AppliedRoute path="/solvetest" exact component={SolveTest} appProps={appProps} />
         <AppliedRoute path="/results" exact component={Results} appProps={appProps} />
         <AppliedRoute path="/customerMenager" exact component={CustomerMenager} appProps={appProps} />
+        <AppliedRoute path="/addCandidate" exact component={AddCandidate} appProps={appProps} />
+        <AppliedRoute path="/confirmationCode" exact component={ConfirmationCode} appProps={appProps} />
         { /* Finally, catch all unmatched routes */ }
         <Route component={NotFound} />
       </Switch>
